pyproject.toml
[project]
name = "ruvds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
]
